requirements.txt
alembic==1.17.2
aniso8601==10.0.1
black==24.10.0
blinker==1.9.0
cfgv==3.5.0
click==8.3.1
coverage==7.13.0
distlib==0.4.0
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.1
filelock==3.20.1
flake8==7.1.1
Flask==3.1.2
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
gunicorn==23.0.0
identify==2.6.15
idna==3.11
iniconfig==2.3.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
mypy==1.11.2
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.0.1
psycopg2-binary==2.9.11
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.1
pytest==8.3.4
pytest-cov==5.0.0
python-dotenv==1.2.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
six==1.17.0
SQLAlchemy==2.0.45
text-unidecode==1.3
tomli==2.3.0
typing_extensions==4.15.0
virtualenv==20.35.4
Werkzeug==3.1.4
